Move Gotway files into separate package
