+++ app/src/main/java/com/cooper/wheellog/WheelData.java
@@ -9,7 +9,7 @@ import android.media.AudioManager;
 
 import com.cooper.wheellog.utils.*;
 import com.cooper.wheellog.utils.Constants.WHEEL_TYPE;
-import com.cooper.wheellog.utils.GotwayAdapter;
+import com.cooper.wheellog.utils.gotway.GotwayAdapter;
 
 import org.json.JSONArray;
 import org.json.JSONException;
