+++ app/src/main/java/com/cooper/wheellog/settings/WheelScreen.kt
@@ -12,6 +12,7 @@ import com.cooper.wheellog.R
 import com.cooper.wheellog.WheelData
 import com.cooper.wheellog.WheelLog.Companion.AppConfig
 import com.cooper.wheellog.utils.*
+import com.cooper.wheellog.utils.gotway.GotwayAdapter
 
 @Composable
 fun wheelScreen()
