+++ app/src/main/java/com/cooper/wheellog/utils/gotway/GotwayUnpacker.java
@@ -1,4 +1,4 @@
-package com.cooper.wheellog.utils;
+package com.cooper.wheellog.utils.gotway;
 
 import java.io.ByteArrayOutputStream;
