+++ app/src/test/java/com/cooper/wheellog/utils/GotwayAdapterTest.kt
@@ -5,6 +5,8 @@ import com.cooper.wheellog.AppConfig
 import com.cooper.wheellog.WheelData
 import com.cooper.wheellog.WheelLog
 import com.cooper.wheellog.utils.Utils.Companion.hexToByteArray
+import com.cooper.wheellog.utils.gotway.GotwayAdapter
+import com.cooper.wheellog.utils.gotway.GotwayUnpacker
 import com.google.common.truth.Truth.assertThat
 import io.mockk.every
 import io.mockk.mockkClass
